refactor(lib/errors): move to TypeScript
