--- src/lib/errors/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-module.exports = {
-  StepsValidationError: require('./steps-validation'),
-  ChunksValidationError: require('./chunks-validation'),
-  SpaceAccessError: require('./space-access'),
-  PayloadValidationError: require('./payload-validation')
-
-};
