+++ src/lib/errors/index.ts
@@ -1,0 +1,11 @@
+import StepsValidationError from './steps-validation'
+import ChunksValidationError from './chunks-validation'
+import SpaceAccessError from './space-access'
+import PayloadValidationError from './payload-validation'
+
+export {
+  StepsValidationError,
+  ChunksValidationError,
+  SpaceAccessError,
+  PayloadValidationError
+}
